MODLD-1069 - code refactoring related put usage for language and bookformat
