+++ src/test/java/org/folio/linked/data/e2e/mappings/hub/language/HubLanguageIT.java
@@ -1,6 +1,9 @@
 package org.folio.linked.data.e2e.mappings.hub.language;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.folio.linked.data.test.TestUtil.getFirstOutgoingResource;
+import static org.folio.linked.data.test.TestUtil.getProperty;
+import static org.folio.linked.data.test.TestUtil.validateResourceType;
 import static org.hamcrest.Matchers.hasItem;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
