+++ src/test/java/org/folio/linked/data/e2e/mappings/instance/bookformat/BookFormatIT.java
@@ -2,12 +2,15 @@ package org.folio.linked.data.e2e.mappings.instance.bookformat;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.folio.linked.data.test.TestUtil.STANDALONE_TEST_PROFILE;
+import static org.folio.linked.data.test.TestUtil.getFirstOutgoingResource;
+import static org.folio.linked.data.test.TestUtil.getProperty;
+import static org.folio.linked.data.test.TestUtil.validateResourceType;
 import static org.folio.linked.data.util.Constants.STANDALONE_PROFILE;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 
 import lombok.SneakyThrows;
 import org.folio.linked.data.e2e.base.IntegrationTest;
-import org.folio.linked.data.e2e.mappings.PutResourceIT;
+import org.folio.linked.data.e2e.mappings.PostResourceIT;
 import org.folio.linked.data.model.entity.Resource;
 import org.junit.jupiter.api.BeforeEach;
 import org.springframework.test.context.ActiveProfiles;
@@ -15,7 +18,7 @@ import org.springframework.test.web.servlet.ResultActions;
 
 @IntegrationTest
 @ActiveProfiles({STANDALONE_PROFILE, STANDALONE_TEST_PROFILE})
-class BookFormatIT extends PutResourceIT {
+class BookFormatIT extends PostResourceIT {
 
   @BeforeEach
   void createWork() {
@@ -70,53 +73,6 @@ class BookFormatIT extends PutResourceIT {
   }
 
   @Override
-  protected String putPayload() {
-    return """
-      {
-         "resource":{
-            "http://bibfra.me/vocab/lite/Instance":{
-               "profileId": 3,
-               "http://bibfra.me/vocab/library/title":[
-                  {
-                     "http://bibfra.me/vocab/library/Title":{
-                        "http://bibfra.me/vocab/library/mainTitle":[ "%s" ]
-                     }
-                  }
-               ],
-               "http://bibfra.me/vocab/library/bookFormat":[
-                  {
-                     "http://bibfra.me/vocab/library/term":[ "8vo" ],
-                     "http://bibfra.me/vocab/lite/link": ["http://id.loc.gov/vocabulary/bookformat/8vo"]
-                  }, {
-                     "http://bibfra.me/vocab/library/term":[ "updated-non-standard" ]
-                  }
-               ],
-               "_workReference": [ { "id": "%s" } ]
-            }
-         }
-      }"""
-      .formatted("TEST: " + this.getClass().getSimpleName(), savedWorkId);
-  }
-
-  @Override
-  @SneakyThrows
-  protected void validateUpdatedApiResponse(ResultActions apiResponse) {
-    var bookFormatPath = "$.resource['http://bibfra.me/vocab/lite/Instance']['http://bibfra.me/vocab/library/bookFormat']";
-    apiResponse
-      .andExpect(
-        jsonPath(bookFormatPath + "[0]['http://bibfra.me/vocab/library/term'][0]")
-          .value("updated-non-standard"))
-      .andExpect(jsonPath(bookFormatPath + "[1]['http://bibfra.me/vocab/library/term'][0]")
-        .value("8vo"))
-      .andExpect(
-        jsonPath(bookFormatPath + "[1]['http://bibfra.me/vocab/library/code'][0]")
-          .value("8vo"))
-      .andExpect(
-        jsonPath(bookFormatPath + "[1]['http://bibfra.me/vocab/lite/link'][0]")
-          .value("http://id.loc.gov/vocabulary/bookformat/8vo"));
-  }
-
-  @Override
   protected void validateGraph(Resource instance) {
     final var expectedBookFormatId = -3739221851083272823L;
     final var expectedCategorySetId = -5037749211942465056L;
@@ -134,27 +90,6 @@ class BookFormatIT extends PutResourceIT {
 
     var categorySet = getFirstOutgoingResource(bookFormat, "http://bibfra.me/vocab/lite/isDefinedBy");
     assertThat(categorySet.getId()).isEqualTo(expectedCategorySetId);
-    validateResourceType(categorySet, "http://bibfra.me/vocab/lite/CategorySet");
-    assertThat(getProperty(categorySet, "http://bibfra.me/vocab/lite/label")).isEqualTo("Book Format");
-    assertThat(getProperty(categorySet, "http://bibfra.me/vocab/lite/link"))
-      .isEqualTo("http://id.loc.gov/vocabulary/bookformat");
-    assertThat(categorySet.getLabel()).isEqualTo("Book Format");
-  }
-
-  @Override
-  protected void validateUpdatedGraph(Resource instance) {
-    assertThat(getProperty(instance, "http://bibfra.me/vocab/library/bookFormat"))
-      .isEqualTo("updated-non-standard");
-
-    var bookFormat = getFirstOutgoingResource(instance, "http://bibfra.me/vocab/library/bookFormat");
-    validateResourceType(bookFormat, "http://bibfra.me/vocab/lite/Category");
-    assertThat(getProperty(bookFormat, "http://bibfra.me/vocab/library/term")).isEqualTo("8vo");
-    assertThat(getProperty(bookFormat, "http://bibfra.me/vocab/library/code")).isEqualTo("8vo");
-    assertThat(getProperty(bookFormat, "http://bibfra.me/vocab/lite/link"))
-      .isEqualTo("http://id.loc.gov/vocabulary/bookformat/8vo");
-    assertThat(bookFormat.getLabel()).isEqualTo("8vo");
-
-    var categorySet = getFirstOutgoingResource(bookFormat, "http://bibfra.me/vocab/lite/isDefinedBy");
     validateResourceType(categorySet, "http://bibfra.me/vocab/lite/CategorySet");
     assertThat(getProperty(categorySet, "http://bibfra.me/vocab/lite/label")).isEqualTo("Book Format");
     assertThat(getProperty(categorySet, "http://bibfra.me/vocab/lite/link"))
