+++ src/test/java/org/folio/linked/data/e2e/mappings/instance/publicationfrequency/PublicationFrequencyIT.java
@@ -2,6 +2,9 @@ package org.folio.linked.data.e2e.mappings.instance.publicationfrequency;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.folio.linked.data.test.TestUtil.TEST_JSON_MAPPER;
+import static org.folio.linked.data.test.TestUtil.getFirstOutgoingResource;
+import static org.folio.linked.data.test.TestUtil.getProperty;
+import static org.folio.linked.data.test.TestUtil.validateResourceType;
 
 import lombok.SneakyThrows;
 import org.folio.linked.data.e2e.mappings.PostResourceIT;
