+++ src/test/java/org/folio/linked/data/e2e/mappings/title/AbstractVariantTitleIT.java
@@ -1,6 +1,8 @@
 package org.folio.linked.data.e2e.mappings.title;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.folio.linked.data.test.TestUtil.getOutgoingResources;
+import static org.folio.linked.data.test.TestUtil.getProperty;
 import static org.hamcrest.Matchers.hasItem;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
