+++ src/test/java/org/folio/linked/data/e2e/mappings/work/dissertation/WorkDissertationIT.java
@@ -3,6 +3,9 @@ package org.folio.linked.data.e2e.mappings.work.dissertation;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.tuple;
 import static org.folio.linked.data.test.TestUtil.TEST_JSON_MAPPER;
+import static org.folio.linked.data.test.TestUtil.getOutgoingResources;
+import static org.folio.linked.data.test.TestUtil.getProperty;
+import static org.folio.linked.data.test.TestUtil.validateResourceType;
 
 import java.util.Comparator;
 import java.util.List;
