+++ src/test/java/org/folio/linked/data/e2e/mappings/work/hub/AbstractHubReferenceRelationIT.java
@@ -3,6 +3,9 @@ package org.folio.linked.data.e2e.mappings.work.hub;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.folio.ld.dictionary.ResourceTypeDictionary.HUB;
 import static org.folio.linked.data.test.TestUtil.TEST_JSON_MAPPER;
+import static org.folio.linked.data.test.TestUtil.getFirstOutgoingResource;
+import static org.folio.linked.data.test.TestUtil.getProperty;
+import static org.folio.linked.data.test.TestUtil.validateResourceType;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 
 import lombok.SneakyThrows;
