+++ src/test/java/org/folio/linked/data/e2e/mappings/work/note/WorkNotesIT.java
@@ -2,6 +2,8 @@ package org.folio.linked.data.e2e.mappings.work.note;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.folio.linked.data.test.TestUtil.TEST_JSON_MAPPER;
+import static org.folio.linked.data.test.TestUtil.getProperties;
+import static org.folio.linked.data.test.TestUtil.getProperty;
 
 import java.util.HashMap;
 import java.util.HashSet;
